chat_URL/requirements.txt
numpy==1.23.5
pandas==1.5.3
streamlit==1.24.0
matplotlib==3.6.3
seaborn==0.12.2
scikit-learn==1.2.1
requests==2.25.1
urllib3==1.26.5
pyrebase4==4.3.0
six==1.15.0
python-dateutil==2.8.2
pymongo==4.3.3
requests-toolbelt==0.9.1
openai==0.27.6
sentence-transformers==2.2.2
firecrawl==0.0.20
langchain==0.0.159
